adding `cosmosDBTrigger` to the list of builtin providers
