+++ test/WebJobs.Script.Tests.Integration/NodeEndToEndTests.cs
@@ -35,12 +35,6 @@ namespace Microsoft.Azure.WebJobs.Script.Tests
         }
 
         [Fact]
-        public async Task CosmosDBTriggerToBlobTest()
-        {
-            await CosmosDBTriggerToBlobTestImpl();
-        }
-
-        [Fact]
         public async Task BlobTriggerToBlobTest()
         {
             TestHelpers.ClearFunctionLogs("BlobTriggerToBlob");
